Optimized getLatexFreeField a bit.
Now, it converts the field name to lower case only once.

The earlier version also had a bug that it wasn't converting the name to lower
case before storing it in the map. So, if the same field was called many times
with different capitalizations (title, Title, TITLE etc.), we would not use
the cached value, and also create a new entry in the cache every time.